+++ packages/core/CHANGELOG.md
@@ -7,7 +7,7 @@
 ### Features
 
 - add `createUnrefCalls`
-- add `useSet`, `useMap`, `useWeakSet`, `useWeakMap` hooks
+- add `useRefSet`, `useRefMap`, `useRefWeakSet`, `useRefWeakMap` hooks
 
 ### Changes